Allow users to override the extension passed to vale checker (#43)
* feat: Allow users to override the extension passed to vale checker

Exposes an option, `flymake-collection-vale-extension-function`, that is a
function that returns the extension (value of the --ext flag passed to vale) of
the current file. The default value, the function
`flymake-collection-vale-default-extension-function`, preserves the previous
behavior of the vale checker.

* chore: Add workflow_dispatch option

* fix: Pass flymake-collection-source as an argument

Previously flymake-collection-vale-default-extension-function tried to access
the value of flymake-collection-source directly. This is erroneous since that
variable is only accessible within the form defining the checker.

To fix this, we let flymake-collection-vale-default-extension-function accept a
buffer as an argument and call flymake-collection-vale-extension-function with
flymake-collection-source as the argument.

---------

Co-authored-by: Mohsin Kaleem <mohkale@kisara.moe>